添加代码历史查看功能 后端：静态文件资源添加版本 前端：新增历史代码对比页

+++ src/main/java/com/rong/rongcodemother/model/vo/AppVO.java
@@ -32,7 +32,7 @@ public class AppVO implements Serializable {
     /**
      * 应用版本
      */
-    private int appVersion;
+    private Short appVersion;
 
     /**
      * 代码生成类型（枚举）
